pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Vertex AI Agent Builder"
requires-python = ">=3.11,<3.12"
dependencies = [
    "cloudpickle>=3.1.2",
    "google-auth",
    "google-cloud-aiplatform>=1.72.0",
    "google-api-core[grpc]>=2.0.0",
    "protobuf>=4.25.0,<5.0.0",
    "grpcio>=1.60.0",
    "langchain>=1.1.0",
    "langchain-google-vertexai>=3.1.1",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "googleapis-common-protos[grpc]>=1.72.0",
    "sift-stack-py>=0.9.1",
]
